+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.1](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.9.0...v4.9.1) (2023-01-27)
+
+### Bug Fixes
+
+- **sendcloud:** email and shippingmethod to sendcloud ([8dd3d30](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/8dd3d306ad0382838a198e0932fe3719ec5238ca))
+
+## [4.9.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.8.0...v4.9.0) (2023-01-27)
+
+### Features
+
+- **shipping:** variant weight overrides ([dbc2f98](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/dbc2f98cc51af91b7dd2138d40e4d61af8e24c20))
+- **variant-bulk-update:** added test workflow ([7194e4a](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/7194e4a68d3eb2861d75859b70b16abf3a7e0d90))
+- **variant-bulk-update:** finished plugin ([0623601](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/06236018fe5420e58310faf13f8b1a0d70305ff1))
+- **variant-bulk-update:** readme update ([c6ea019](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/c6ea019e00b8a3a59a7a6ad59d8e679dee287cb5))
+- **variant-bulk-update:** readme updates ([23321ee](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/23321eef5931800fc15717ed016753ae24038fd9))
+- wip ([b6b9b3e](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/b6b9b3e871bee715c69bdb5c713a39fd26364f06))
+
 ## [4.8.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.7.2...v4.8.0) (2023-01-24)
 
 ### Features